# backend/docs/openapi-admin.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MBGA Admin API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/admin/auth/otp/request": {
      "post": {
        "tags": [
          "Admin Authentication"
        ],
        "summary": "Request Otp",
        "operationId": "request_otp_api_v1_admin_auth_otp_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OTPRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OTPRequestResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error. `detail.fields` lists the invalid fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests. See the Retry-After header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auth/otp/resend": {
      "post": {
        "tags": [
          "Admin Authentication"
        ],
        "summary": "Resend Otp",
        "operationId": "resend_otp_api_v1_admin_auth_otp_resend_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OTPResendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OTPRequestResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request state (for example an incorrect or used code).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The account may not perform this action or use this app.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error. `detail.fields` lists the invalid fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests. See the Retry-After header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auth/otp/verify": {
      "post": {
        "tags": [
          "Admin Authentication"
        ],
        "summary": "Verify Otp",
        "operationId": "verify_otp_api_v1_admin_auth_otp_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OTPVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OTPVerifyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request state (for example an incorrect or used code).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The account may not perform this action or use this app.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error. `detail.fields` lists the invalid fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auth/token/refresh": {
      "post": {
        "tags": [
          "Admin Authentication"
        ],
        "summary": "Refresh Token",
        "operationId": "refresh_token_api_v1_admin_auth_token_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or revoked session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The account may not perform this action or use this app.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error. `detail.fields` lists the invalid fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auth/logout": {
      "post": {
        "tags": [
          "Admin Authentication"
        ],
        "summary": "Logout",
        "description": "Revokes the session of the given refresh token. Always returns 204, even when nothing was revoked.",
        "operationId": "logout_api_v1_admin_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation error. `detail.fields` lists the invalid fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auth/logout-all": {
      "post": {
        "tags": [
          "Admin Authentication"
        ],
        "summary": "Logout All",
        "operationId": "logout_all_api_v1_admin_auth_logout_all_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing, invalid, expired or revoked session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The account may not perform this action or use this app.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/auth/me": {
      "get": {
        "tags": [
          "Admin Authentication"
        ],
        "summary": "Me",
        "operationId": "me_api_v1_admin_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or revoked session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The account may not perform this action or use this app.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CurrentUserResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "login_channel": {
            "$ref": "#/components/schemas/LoginChannel"
          },
          "user_type": {
            "type": "string",
            "title": "User Type",
            "description": "CUSTOMER, DELIVERY_PARTNER, MERCHANT_STAFF or ADMIN."
          },
          "session_type": {
            "type": "string",
            "title": "Session Type",
            "description": "`access` (full app session) or `onboarding` (customer registration only)."
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role",
            "description": "Primary role code for this app, e.g. manager, driver, customer."
          },
          "active_roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Active Roles",
            "description": "Distinct role codes allowed on this app."
          },
          "account_status": {
            "type": "string",
            "title": "Account Status"
          },
          "approval_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Status",
            "description": "Merchant, delivery-profile or customer approval state."
          },
          "profile_completion_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Completion Status",
            "description": "Customer only: NOT_STARTED, INCOMPLETE or COMPLETE."
          },
          "next_action": {
            "$ref": "#/components/schemas/NextAction"
          },
          "merchant": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MerchantSummary"
              },
              {
                "type": "null"
              }
            ],
            "description": "Merchant and delivery apps only."
          },
          "delivery_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeliveryProfileSummary"
              },
              {
                "type": "null"
              }
            ],
            "description": "Delivery app only."
          },
          "customer_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CustomerProfileSummary"
              },
              {
                "type": "null"
              }
            ],
            "description": "Customer app only, once details are submitted."
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "mobile_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mobile Number"
          },
          "country_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Code"
          },
          "effective_permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Effective Permissions"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Deprecated: same value as account_status.",
            "deprecated": true
          },
          "profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeliveryProfileSummary"
              },
              {
                "$ref": "#/components/schemas/CustomerProfileSummary"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile",
            "description": "Deprecated: use delivery_profile or customer_profile.",
            "deprecated": true
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "login_channel",
          "user_type",
          "session_type",
          "account_status",
          "next_action",
          "status"
        ],
        "title": "CurrentUserResponse"
      },
...
      "DeviceInfo": {
        "properties": {
          "device_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id",
            "description": "Stable install ID generated by the app."
          },
          "device_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Type",
            "description": "android or ios."
          },
          "device_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Name"
          },
          "app_version": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40
              },
              {
                "type": "null"
              }
            ],
            "title": "App Version"
          },
          "fcm_token": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Fcm Token",
            "description": "Push token. Stored on the session created by OTP verification or refresh and cleared on logout."
          }
        },
        "type": "object",
        "title": "DeviceInfo"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable machine-readable error code."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Safe, user-facing message."
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array",
            "title": "Fields"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "description": "Also returned in the X-Request-ID header."
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
...
      "LoginChannel": {
        "type": "string",
        "enum": [
          "CUSTOMER",
          "DELIVERY",
          "MERCHANT",
          "ADMIN"
        ],
        "title": "LoginChannel"
      },
      "LogoutRequest": {
        "properties": {
          "refresh_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Token"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id",
            "description": "Ignored.",
            "deprecated": true
          }
        },
        "type": "object",
        "title": "LogoutRequest"
      },
...
      "MerchantSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "approval_status": {
            "type": "string",
            "title": "Approval Status"
          },
          "staff_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Staff Type"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "status",
          "approval_status"
        ],
        "title": "MerchantSummary"
      },
...
      "NextAction": {
        "type": "string",
        "enum": [
          "COMPLETE_PROFILE",
          "SUBMIT_DOCUMENTS",
          "WAIT_FOR_APPROVAL",
          "SIGN_IN",
          "CONTACT_SUPPORT",
          "OPEN_CUSTOMER_HOME",
          "OPEN_DELIVERY_HOME",
          "OPEN_MERCHANT_HOME",
          "OPEN_ADMIN_DASHBOARD"
        ],
        "title": "NextAction"
      },
      "OTPRequest": {
        "properties": {
          "mobile_number": {
            "type": "string",
            "maxLength": 15,
            "minLength": 6,
            "title": "Mobile Number"
          },
          "country_code": {
            "type": "string",
            "maxLength": 5,
            "minLength": 1,
            "title": "Country Code",
            "default": "+91"
          },
          "login_channel": {
            "$ref": "#/components/schemas/LoginChannel",
            "description": "Ignored: the app channel is taken from the URL.",
            "default": "CUSTOMER",
            "deprecated": true
          },
          "device": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeviceInfo"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "mobile_number"
        ],
        "title": "OTPRequest"
      },
      "OTPRequestResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "OTP request accepted"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "resend_after": {
            "type": "integer",
            "title": "Resend After"
          },
          "code": {
            "type": "string",
            "title": "Code",
            "default": "OTP_REQUEST_ACCEPTED"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "expires_in",
          "resend_after"
        ],
        "title": "OTPRequestResponse"
      },
      "OTPResendRequest": {
        "properties": {
          "request_id": {
            "type": "string",
            "maxLength": 36,
            "minLength": 1,
            "title": "Request Id"
          },
          "device": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeviceInfo"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "OTPResendRequest"
      },
      "OTPVerifyRequest": {
        "properties": {
          "request_id": {
            "type": "string",
            "maxLength": 36,
            "minLength": 1,
            "title": "Request Id"
          },
          "mobile_number": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 15,
                "minLength": 6
              },
              {
                "type": "null"
              }
            ],
            "title": "Mobile Number",
            "description": "Ignored.",
            "deprecated": true
          },
          "country_code": {
            "type": "string",
            "maxLength": 5,
            "minLength": 1,
            "title": "Country Code",
            "description": "Ignored.",
            "default": "+91",
            "deprecated": true
          },
          "login_channel": {
            "$ref": "#/components/schemas/LoginChannel",
            "description": "Ignored.",
            "default": "CUSTOMER",
            "deprecated": true
          },
          "otp": {
            "type": "string",
            "pattern": "^[0-9]{4}$",
            "title": "Otp",
            "description": "Numeric code with OTP_LENGTH digits (4 in the current configuration)."
          },
          "device": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeviceInfo"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "otp"
        ],
        "title": "OTPVerifyRequest"
      },
      "OTPVerifyResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "login_channel": {
            "$ref": "#/components/schemas/LoginChannel"
          },
          "user_type": {
            "type": "string",
            "title": "User Type",
            "description": "CUSTOMER, DELIVERY_PARTNER, MERCHANT_STAFF or ADMIN."
          },
          "session_type": {
            "type": "string",
            "title": "Session Type",
            "description": "`access` (full app session) or `onboarding` (customer registration only)."
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role",
            "description": "Primary role code for this app, e.g. manager, driver, customer."
          },
          "active_roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Active Roles",
            "description": "Distinct role codes allowed on this app."
          },
          "account_status": {
            "type": "string",
            "title": "Account Status"
          },
          "approval_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Status",
            "description": "Merchant, delivery-profile or customer approval state."
          },
          "profile_completion_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Completion Status",
            "description": "Customer only: NOT_STARTED, INCOMPLETE or COMPLETE."
          },
          "next_action": {
            "$ref": "#/components/schemas/NextAction"
          },
          "merchant": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MerchantSummary"
              },
              {
                "type": "null"
              }
            ],
            "description": "Merchant and delivery apps only."
          },
          "delivery_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeliveryProfileSummary"
              },
              {
                "type": "null"
              }
            ],
            "description": "Delivery app only."
          },
          "customer_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CustomerProfileSummary"
              },
              {
                "type": "null"
              }
            ],
            "description": "Customer app only, once details are submitted."
          },
          "token": {
            "$ref": "#/components/schemas/TokenPair"
          },
          "is_new_user": {
            "type": "boolean",
            "title": "Is New User",
            "description": "True when this verification created the account.",
            "default": false
          },
          "code": {
            "type": "string",
            "title": "Code",
            "default": "OTP_VERIFIED"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "OTP verified"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "login_channel",
          "user_type",
          "session_type",
          "account_status",
          "next_action",
          "token"
        ],
        "title": "OTPVerifyResponse"
      },
...
      "RefreshTokenRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "minLength": 1,
            "title": "Refresh Token"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id",
            "description": "Deprecated: use device.device_id.",
            "deprecated": true
          },
          "device": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeviceInfo"
              },
              {
                "type": "null"
              }
            ],
            "description": "Send when the push token or app version changed; omitted values are kept."
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshTokenRequest"
      },
...
      "TokenPair": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "$ref": "#/components/schemas/TokenType",
            "default": "Bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "description": "Access-token lifetime in seconds."
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "expires_in"
        ],
        "title": "TokenPair"
      },
      "TokenType": {
        "type": "string",
        "enum": [
          "Bearer"
        ],
        "title": "TokenType"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}